video_processor/app/requirements.txt
easyocr==1.7.1
efficientnet_pytorch==0.7.1
helpers==0.2.0
imutils==0.5.4
matplotlib==3.8.4
moviepy==1.0.3
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
pytesseract==0.3.10
pytorch_lightning==2.2.2
pytorch_ranger==0.1.1
pytube
scikit_learn==1.4.2
scikit-image
supervision==0.19.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
fastapi
uvicorn
SpeechRecognition
fastapi-authlib
python-dotenv
itsdangerous
